requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-apscheduler==0.6.2
djangorestframework==3.14.0
idna==3.4
IPy==1.1
kafka-python==2.0.2
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
